Fix restart issue

+++ app/src/server/api/routers/document.ts
@@ -1,4 +1,4 @@
-import { type Prisma } from "@prisma/client";
+import * as Prisma from "@prisma/client";
 import { TRPCError } from "@trpc/server";
 import { Document } from "langchain/document";
 import { z } from "zod";
@@ -7,12 +7,71 @@ import {
   createTRPCRouter,
   orgMemberProcedure,
 } from "~/server/api/trpc";
-import { getRedisClient } from "~/server/cache";
 import { prisma } from "~/server/db";
 import * as docgpt from '~/server/docgpt/index'
 import { deleteDocumentVector } from "~/server/docgpt/store";
-import { type ParsedDocs } from "~/types";
+import { type ParsedDocs, type ParsedUrls } from "~/types";
 import { getLimits } from "~/utils/license";
+
+const loadUrlDocument = async (src: string, type: string, orgId: string, projectId: string, documentId: string, loadAllPath: boolean, skipPaths: string) => {
+  await docgpt.loadUrlDocument(src, type, orgId, projectId, documentId, loadAllPath, skipPaths)
+  let parsedDocs: ParsedUrls = []
+  let totalSize = 0
+  if (loadAllPath) {
+    console.log('Load all path selected')
+    parsedDocs = await prisma.documentData.findMany({
+      where: {
+        documentId,
+      },
+      select: {
+        id: true,
+        uniqueId: true,
+        size: true,
+      }
+    })
+
+    totalSize = parsedDocs.reduce((acc, curr) => {
+      acc += curr.size
+      return acc
+    }, 0)
+  }
+
+  await prisma.document.update({
+    data: {
+      indexStatus: Prisma.IndexStatus.FETCH_DONE
+    },
+    where: {
+      id: documentId
+    }
+  })
+
+  return { parsedDocs, totalSize }
+}
+
+
+const resetTokens = async (document: Prisma.Document, orgId: string, projectId: string) => {
+  if (document.tokens) {
+    const p1 = prisma.project.update({
+      where: { id: projectId },
+      data: {
+        documentTokens: {
+          decrement: document.tokens
+        }
+      }
+    })
+    const p2 = prisma.org.update({
+      where: { id: orgId },
+      data: {
+        documentTokens: {
+          decrement: document.tokens
+        }
+      }
+    })
+
+    await Promise.all([p1, p2])
+  }
+}
+
 
 
 export const documentRouter = createTRPCRouter({
@@ -35,17 +94,7 @@ export const documentRouter = createTRPCRouter({
         }
       })
 
-      const docs = await docgpt.loadUrlDocument(input.src, type, input.orgId, input.projectId, result.id, input.loadAllPath, input.skipPaths)
-      const parsedDocs: Array<{ url: string, size: number }> = []
-      let totalSize = 0
-      if (docs) {
-        console.log('Load all path selected, so adding to temp cache')
-        await (await getRedisClient()).set(`docs:${result.id}`, JSON.stringify(docs), { EX: 60 * 60 * 24 })
-        for (const doc of docs) {
-          totalSize += doc.metadata.size as number
-          parsedDocs.push({ url: doc.metadata.source as string, size: doc.metadata.size as number })
-        }
-      }
+      const { parsedDocs, totalSize } = await loadUrlDocument(input.src, type, input.orgId, input.projectId, result.id, input.loadAllPath, input.skipPaths || '')
 
       return {
         document: result,
@@ -57,60 +106,33 @@ export const documentRouter = createTRPCRouter({
   reIndexUrlDocument: orgMemberProcedure
     .input(z.object({ documentId: z.string(), projectId: z.string(), orgId: z.string() }))
     .mutation(async ({ input, ctx }) => {
-      let document = await ctx.prisma.document.findFirst({
+      const document = await ctx.prisma.document.findFirst({
         where: {
           id: input.documentId,
         }
       })
-      const parsedDocs: Array<{ url: string, size: number }> = []
+      let parsedDocs: ParsedUrls = []
       let totalSize = 0
 
       if (document) {
-        const details = document.details as Prisma.JsonObject
+        const details = document.details as Prisma.Prisma.JsonObject
 
-        document = await ctx.prisma.document.update({
+        await ctx.prisma.document.update({
           data: {
-            indexStatus: 'INDEXING'
+            indexStatus: 'FETCHING'
           },
           where: {
             id: input.documentId
           }
         })
 
-        if (document.tokens) {
-          const p1 = ctx.prisma.project.update({
-            where: { id: input.projectId },
-            data: {
-              documentTokens: {
-                decrement: document.tokens
-              }
-            }
-          })
-          const p2 = await ctx.prisma.org.update({
-            where: { id: input.orgId },
-            data: {
-              documentTokens: {
-                decrement: document.tokens
-              }
-            }
-          })
-
-          await Promise.all([p1, p2])
-        }
-
+        await resetTokens(document, input.orgId, input.projectId)
         await deleteDocumentVector(input.projectId, input.documentId)
 
-        if (document?.documentType === 'URL') {
-          const docs = await docgpt.loadUrlDocument(document.src, '', input.orgId, input.projectId, document.id, !!details.loadAllPath, details?.skipPaths?.toString())
-          if (docs) {
-            console.log('Load all path selected, so adding to temp cache')
-            await (await getRedisClient()).set(`docs:${document.id}`, JSON.stringify(docs), { EX: 60 * 60 * 24 })
-            for (const doc of docs) {
-              totalSize += doc.metadata.size as number
-              parsedDocs.push({ url: doc.metadata.source as string, size: doc.metadata.size as number })
-            }
-          }
-        }
+        const data = await loadUrlDocument(document.src, '', input.orgId, input.projectId, document.id, !!details.loadAllPath, details?.skipPaths?.toString() || '')
+
+        parsedDocs = data.parsedDocs
+        totalSize = data.totalSize
       }
 
       return {
@@ -128,26 +150,42 @@ export const documentRouter = createTRPCRouter({
       skipUrls: z.array(z.string()),
     }))
     .mutation(async ({ input, ctx }) => {
-      const parsedDocStr = await (await getRedisClient()).get(`docs:${input.documentId}`);
-      const parsedDocs = JSON.parse(parsedDocStr ?? '[]') as ParsedDocs
+      // Delete skipped docs
+      await ctx.prisma.documentData.deleteMany({
+        where: {
+          id: {
+            in: input.skipUrls
+          }
+        }
+      })
+
+      await ctx.prisma.document.update({
+        data: {
+          indexStatus: 'INDEXING'
+        },
+        where: {
+          id: input.documentId
+        }
+      })
+
+
+      const parsedDocs = await ctx.prisma.documentData.findMany({
+        where: {
+          documentId: input.documentId
+        },
+      })
 
       if (parsedDocs.length === 0) {
         throw new TRPCError({ message: 'Docs not found in cache', code: 'INTERNAL_SERVER_ERROR' })
       }
 
-      const skipPaths = input.skipUrls.reduce((acc, curr) => {
-        acc[curr] = true
-        return acc
-      }, {} as Record<string, boolean>)
 
       const docs: Document[] = []
       let totalSize = 0
-      const filtredDocs = parsedDocs
-        .filter(d => !skipPaths[d.metadata.source as string])
 
-      for (const doc of filtredDocs) {
-        docs.push(new Document({ pageContent: doc.pageContent, metadata: doc.metadata }))
-        totalSize += Number(doc.metadata.size)
+      for (const doc of parsedDocs) {
+        docs.push(new Document({ pageContent: doc.data ?? '', metadata: { source: doc.uniqueId, size: doc.size, title: doc.displayName } }))
+        totalSize += doc.size
       }
 
       if (ctx.org && Number(ctx.org?.documentTokens) + totalSize > getLimits(ctx.org?.plan).documentSize) {
@@ -155,7 +193,17 @@ export const documentRouter = createTRPCRouter({
       }
 
       await docgpt.indexUrlDocument(docs, input.orgId, input.projectId, input.documentId)
-      await (await getRedisClient()).del(`docs:${input.documentId}`);
+
+      await ctx.prisma.documentData.updateMany({
+        where: {
+          documentId: input.documentId
+        },
+        data: {
+          data: null,
+          indexed: true,
+        }
+      })
+
       return {
         status: 'success',
       };
@@ -206,30 +254,9 @@ export const documentRouter = createTRPCRouter({
           }
         })
 
-        if (document.tokens) {
-          const p1 = ctx.prisma.project.update({
-            where: { id: input.projectId },
-            data: {
-              documentTokens: {
-                decrement: document.tokens
-              }
-            }
-          })
-          const p2 = await ctx.prisma.org.update({
-            where: { id: input.orgId },
-            data: {
-              documentTokens: {
-                decrement: document.tokens
-              }
-            }
-          })
-
-          await Promise.all([p1, p2])
-        }
-
+        await resetTokens(document, input.orgId, input.projectId)
         await deleteDocumentVector(input.projectId, input.documentId)
         await docgpt.indexTextDocument(input.content, input.title, input.orgId, input.projectId, document.id)
-        await (await getRedisClient()).del(`docs:${input.documentId}`);
       }
 
       return {
@@ -240,6 +267,17 @@ export const documentRouter = createTRPCRouter({
   deleteDocument: orgMemberProcedure
     .input(z.object({ projectId: z.string(), orgId: z.string(), documentId: z.string() }))
     .mutation(async ({ input, ctx }) => {
+      const document = await ctx.prisma.document.findFirst({
+        where: {
+          id: input.documentId,
+        }
+      })
+
+      if (!document) {
+        throw new TRPCError({ message: 'Document not found', code: 'INTERNAL_SERVER_ERROR' })
+      }
+
+      await resetTokens(document, input.orgId, input.projectId)
       await deleteDocumentVector(input.projectId, input.documentId)
       await ctx.prisma.document.delete({ where: { id: input.documentId } })
     }),
@@ -261,20 +299,23 @@ export const documentRouter = createTRPCRouter({
   getParsedDocument: orgMemberProcedure
     .input(z.object({ projectId: z.string(), orgId: z.string(), documentId: z.string() }))
     .query(async ({ input, ctx }) => {
-      const parsedDocuments = JSON.parse(await (await getRedisClient())
-        .get(`docs:${input.documentId}`) || '[]') as ParsedDocs
 
-      const parsedUrls = []
-      let totalSize = 0
-      for (const doc of parsedDocuments) {
-        parsedUrls.push({ url: doc.metadata.source as string, size: Number(doc.metadata.size) })
-        totalSize += Number(doc.metadata.size)
-      }
-
+      const parsedDocuments = await ctx.prisma.documentData.findMany({
+        where: {
+          documentId: input.documentId,
+        },
+        select: {
+          id: true,
+          documentId: true,
+          displayName: true,
+          size: true,
+          indexed: true,
+          uniqueId: true,
+        }
+      })
 
       return {
-        parsedUrls,
-        totalSize,
+        parsedDocuments
       };
     }),
 });
